fix: unable to redirect to login page after logout
