+++ IDRIS/src/UserContext.tsx
@@ -1,10 +1,4 @@
-import {
-  createContext,
-  useContext,
-  useState,
-  ReactNode,
-  useMemo,
-} from "react";
+import { createContext, useContext } from "react";
 
 type UserContextType = {
   userId: number | null;
@@ -25,37 +19,11 @@ export const UserContext = createContext<UserContextType>({
   email: "",
   username: "",
   isUserReady: false,
-  setUserId: () => { },
-  setUserType: () => { },
-  setEmail: () => { },
-  setUsername: () => { },
-  setUserReady: () => { },
+  setUserId: () => {},
+  setUserType: () => {},
+  setEmail: () => {},
+  setUsername: () => {},
+  setUserReady: () => {},
 });
 
-export const UserProvider = ({ children }: { children: ReactNode }) => {
-  const [userId, setUserId] = useState<number | null>(null);
-  const [userType, setUserType] = useState("");
-  const [email, setEmail] = useState("");
-  const [username, setUsername] = useState("");
-  const [isUserReady, setUserReady] = useState(false);
-
-  const value = useMemo(
-    () => ({
-      userId,
-      setUserId,
-      userType,
-      setUserType,
-      email,
-      setEmail,
-      username,
-      setUsername,
-      isUserReady,
-      setUserReady,
-    }),
-    [userId, userType, email, username, isUserReady],
-  );
-
-  return <UserContext.Provider value={value}>{children}</UserContext.Provider>;
-};
-
-  export const useUserContext = () => useContext(UserContext);
+export const useUserContext = () => useContext(UserContext);
